Adjust about main process.
move main process util function in `utils/process` path.
delete unused code about `store.js` and `widgetmanager.js`
