+++ app/constants/path/index.js
@@ -10,3 +10,5 @@ export const WIDGET_BASIC_PATH = process.env.NODE_ENV === 'development' ? 'app/p
 export const PREFERENCE_BASIC_PATH = process.env.NODE_ENV === 'development' ? 'app/page/setting' : 'build';
 export const WIDGET_PATH = `${ROOT_PATH}/${WIDGET_BASIC_PATH}/widget.html`;
 export const PREFERENCE_PATH = `${ROOT_PATH}/${PREFERENCE_BASIC_PATH}//index.html`;
+
+export const TRAY_ICON_PATH = `${ROOT_PATH}/app/assets/iconTemplate.png`;
