+++ app/utils/__tests__/openAllWidgetStatusOpen.spec.js
@@ -2,8 +2,8 @@ import { BrowserWindow } from 'electron';
 import Immutable from 'immutable';
 import store from 'store/storeMain';
 import * as statusActions from 'actions/status';
-import * as makeWidget from 'utils/makeWidgetWindow';
-import openAllWidgetStatusOpen from '../openAllWidgetStatusOpen';
+import * as makeWidget from 'utils/process/makeWidgetWindow';
+import openAllWidgetStatusOpen from '../process/openAllWidgetStatusOpen';
 
 describe('test openAllWidgetStatusOpen', () => {
   it('should call openAllWidgetStatusOpen', () => {
