+++ app/utils/process/updateWidgetContentBounds.js
